Refactored Code Ensure all references to SafariWebAuth switch based upon iOS version Added backwards compatibility Xcode 8 Added pod exclusions Updated tests
