+++ Auth0.podspec
@@ -8,8 +8,12 @@ web_auth_files = [
   'Auth0/TransactionStore.swift',
   'Auth0/WebAuth.swift',
   'Auth0/WebAuthError.swift',
+  'Auth0/SafariWebAuthenticatable.swift',
   'Auth0/SafariWebAuth.swift',
+  'Auth0/SafariWebAuthSession.swift',
+  'Auth0/AuthSession.swift',
   'Auth0/SafariSession.swift',
+  'Auth0/SafariAuthenticationSession.swift',
   'Auth0/SilentSafariViewController.swift',
   'Auth0/NativeAuth.swift',
   'Auth0/AuthProvider.swift',
