+++ Auth0/SafariSession.swift
@@ -23,63 +23,16 @@
 import UIKit
 import SafariServices
 
-class SafariSession: NSObject, AuthTransaction {
+class SafariSession: AuthSession {
 
     typealias FinishSession = (Result<Credentials>) -> Void
 
     weak var controller: UIViewController?
 
-    let redirectURL: URL
-    let state: String?
-    let finish: FinishSession
-    let handler: OAuth2Grant
-    let logger: Logger?
-
     init(controller: SFSafariViewController, redirectURL: URL, state: String? = nil, handler: OAuth2Grant, finish: @escaping FinishSession, logger: Logger?) {
         self.controller = controller
-        self.redirectURL = redirectURL
-        self.state = state
-        self.finish = finish
-        self.handler = handler
-        self.logger = logger
-        super.init()
+        super.init(redirectURL: redirectURL, state: state, handler: handler, finish: finish, logger: logger)
         controller.delegate = self
-    }
-
-    /**
-     Tries to resume (and complete) the OAuth2 session from the received URL
-
-     - parameter url:     url received in application's AppDelegate
-     - parameter options: a dictionary of launch options received from application's AppDelegate
-
-     - returns: `true` if the url completed (successfuly or not) this session, `false` otherwise
-     */
-    func resume(_ url: URL, options: [UIApplicationOpenURLOptionsKey: Any] = [:]) -> Bool {
-        self.logger?.trace(url: url, source: "iOS Safari")
-        guard url.absoluteString.lowercased().hasPrefix(self.redirectURL.absoluteString.lowercased()) else { return false }
-
-        guard
-            let components = URLComponents(url: url, resolvingAgainstBaseURL: true)
-            else {
-                self.finish(.failure(error: AuthenticationError(string: url.absoluteString, statusCode: 200)))
-                return false
-            }
-        var items = self.handler.values(fromComponents: components)
-        guard has(state: self.state, inItems: items) else { return false }
-        if items["error"] != nil {
-            self.finish(.failure(error: AuthenticationError(info: items, statusCode: 0)))
-        } else {
-            self.handler.credentials(from: items, callback: self.finish)
-        }
-        return true
-    }
-
-    func cancel() {
-        self.finish(Result.failure(error: WebAuthError.userCancelled))
-    }
-
-    private func has(state: String?, inItems items: [String: String]) -> Bool {
-        return state == nil || items["state"] == state
     }
 }
