+++ Auth0/_ObjectiveWebAuth.swift
@@ -27,15 +27,31 @@ import UIKit
 // swiftlint:disable:next type_name
 public class _ObjectiveOAuth2: NSObject {
 
-    private(set) var webAuth: SafariWebAuth
+    private(set) var webAuth: SafariWebAuthenticatable
 
     @objc public override init() {
         let values = plistValues(bundle: Bundle.main)!
-        self.webAuth = SafariWebAuth(clientId: values.clientId, url: .a0_url(values.domain))
+        #if swift(>=3.2)
+            if #available(iOS 11.0, *) {
+                self.webAuth = SafariWebAuthSession(clientId: values.clientId, url: .a0_url(values.domain))
+            } else {
+                self.webAuth = SafariWebAuth(clientId: values.clientId, url: .a0_url(values.domain))
+            }
+        #else
+            self.webAuth = SafariWebAuth(clientId: values.clientId, url: .a0_url(values.domain))
+        #endif
     }
 
     @objc public init(clientId: String, url: URL) {
-        self.webAuth = SafariWebAuth(clientId: clientId, url: url)
+        #if swift(>=3.2)
+            if #available(iOS 11.0, *) {
+                self.webAuth = SafariWebAuthSession(clientId: clientId, url: url)
+            } else {
+                self.webAuth = SafariWebAuth(clientId: clientId, url: url)
+            }
+        #else
+            self.webAuth = SafariWebAuth(clientId: clientId, url: url)
+        #endif
     }
 
     @objc public func addParameters(_ parameters: [String: String]) {
